+++ binless/R/CppDocs.R
@@ -6,7 +6,7 @@
 #' of the decay and genomic biases, and one decay and one genomic bias for all datasets. This fast 
 #' model is meant for a quick overview of the data.
 #'
-#' @usage fast_binless(obs, nbins, alpha, lam2, nouter = 25, tol_val = 2e-1,
+#' @usage fast_binless(obs, nbins, alpha, lam2, lam1 = 0, nouter = 25, tol_val = 2e-1,
 #'  bg_steps = 5, free_decay = 10000)
 #'
 #' @param obs DataFrame containing at least 4 named columns: name, bin1, bin2 and observed.
@@ -18,6 +18,7 @@
 #' @param alpha the dispersion of the negative binomial
 #' @param lam2 numeric positive value for the fusion penalty, or a vector of
 #' the same size as the number of datasets.
+#' @param lam1 numeric positive value (or vector) for the significance threshold (default is zero)
 #' @param nouter unsigned The maximum number of iterations that should be performed (default 25)
 #' @param tol_val double tolerance on the values for convergence and the fused lasso (default 2e-1)
 #' @param bg_steps unsigned the maximum number of initial steps where no signal is fitted (default 5)
@@ -31,11 +32,12 @@ NULL
 #' 
 #' Once normalized with \code{\link{fast_binless}}, differences with respect to a reference can be computed.
 #' 
-#' @usage fast_binless_difference(out, ref, alpha, lam2, tol_val = 2e-1)
+#' @usage fast_binless_difference(out, ref, alpha, lam2, lam1 = 0, tol_val = 2e-1)
 #'
 #' @param ref unsigned integer corresponding to the index (starting at 1) of the dataset to use as reference
 #' @param lam2 numeric positive value for the fusion penalty, or a vector of
 #' the same size as the number of datasets minus one.
+#' @param lam1 numeric positive value (or vector of size n-1) for the significance threshold (default is zero)
 #' @inheritParams fast_binless
 #' 
 #' @export