+++ binless/src/csnorm_cpp.cpp
@@ -15,11 +15,10 @@ RCPP_MODULE(binless_cpp) {
     function("wgfl_diff_BIC", &wgfl_diff_BIC, "documentation for wgfl_diff_BIC ");
     
     function("fast_binless", &binless::fast::binless,
-             List::create(_["obs"], _["nbins"], _["lam2"], _["alpha"], _["nouter"]=25, _["tol_val"]=2e-1, _["bg_steps"]=5, _["free_decay"]=10000),
+             List::create(_["obs"], _["nbins"], _["alpha"], _["lam2"], _["lam1"]=0., _["nouter"]=25, _["tol_val"]=2e-1, _["bg_steps"]=5, _["free_decay"]=10000),
              "documentation for fast_binless ");
-    //function("fast_binless_eval_cv", &binless::fast::binless_eval_cv, "documentation for fast_binless_eval_cv ");
     function("fast_binless_difference", &binless::fast::binless_difference,
-             List::create(_["obs"], _["lam2"], _["ref"], _["alpha"], _["tol_val"]=2e-1),
+             List::create(_["obs"], _["ref"], _["alpha"],  _["lam2"], _["lam1"]=0., _["tol_val"]=2e-1),
              "documentation for fast_binless_difference ");
     
     function("generate_spline_base", &generate_spline_base, "documentation for generate_spline_base ");